* merged mini fix from klatexformula-3.2 branch
git-svn-id: https://svn.code.sf.net/p/klatexformula/code/trunk/klatexformula@612 4112663c-bd99-11de-a44c-fbaffaffe1b7

+++ src/klftools/klfutil.cpp
@@ -528,7 +528,7 @@ KLF_EXPORT QByteArray klfSaveVariantToText(const QVariant& value, bool saveListA
     return QByteArray();
 
   // values of value.type() are QMetaType::Type enum entries. See qt's doc.
-  switch (value.type()) {
+  switch ((int)value.type()) {
   case QMetaType::Bool:
     data = value.toBool() ? "true" : "false";
     break;
